+++ sqlserver/tests/test_connection.py
@@ -7,7 +7,7 @@ import mock
 import pytest
 
 from datadog_checks.base import ConfigurationError
-from datadog_checks.sqlserver.connection_manager import ConnectionManager
+from datadog_checks.sqlserver.connection import Connection
 
 pytestmark = pytest.mark.unit
 
@@ -22,7 +22,7 @@ pytestmark = pytest.mark.unit
 )
 def test_will_warn_parameters_for_the_wrong_connection(instance_sql2017_defaults, connector, param):
     instance_sql2017_defaults.update({'connector': connector, param: 'foo'})
-    connection = ConnectionManager({}, instance_sql2017_defaults, None)
+    connection = Connection({}, instance_sql2017_defaults, None)
     connection.log = mock.MagicMock()
     connection._connection_options_validation('somekey', 'somedb')
     connection.log.warning.assert_called_once_with(
@@ -46,7 +46,7 @@ def test_will_warn_parameters_for_the_wrong_connection(instance_sql2017_defaults
 )
 def test_will_fail_for_duplicate_parameters(instance_sql2017_defaults, connector, cs, param):
     instance_sql2017_defaults.update({'connector': connector, param: 'foo', 'connection_string': cs + "=foo"})
-    connection = ConnectionManager({}, instance_sql2017_defaults, None)
+    connection = Connection({}, instance_sql2017_defaults, None)
     match = (
         "%s has been provided both in the connection string and as a configuration option (%s), "
         "please specify it only once" % (cs, param)
@@ -73,7 +73,7 @@ def test_will_fail_for_duplicate_parameters(instance_sql2017_defaults, connector
 def test_will_fail_for_wrong_parameters_in_the_connection_string(instance_sql2017_defaults, connector, cs):
     instance_sql2017_defaults.update({'connector': connector, 'connection_string': cs + '=foo'})
     other_connector = 'odbc' if connector != 'odbc' else 'adodbapi'
-    connection = ConnectionManager({}, instance_sql2017_defaults, None)
+    connection = Connection({}, instance_sql2017_defaults, None)
     match = (
         "%s has been provided in the connection string. "
         "This option is only available for %s connections, however %s has been selected"
